+++ CHANGELOG.md
@@ -7,9 +7,9 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
-## Unreleased
+## 68.2.0
 
-* Adjust cards content elements widths ([PR #](https://github.com/alphagov/govuk_publishing_components/pull/5665))
+* Adjust cards content elements widths ([PR #5665](https://github.com/alphagov/govuk_publishing_components/pull/5665))
 * Use the details component for toggling element visibility in the metadata component ([PR #5592](https://github.com/alphagov/govuk_publishing_components/pull/5592))
 
 ## 68.1.0